DocAgent/requirements.txt


langchain>=0.1.0
openai>=1.0.0
chromadb>=0.4.0
tqdm>=4.65.0
numpy>=1.25.0
python-dotenv>=1.0.0